@@ -4,19 +4,25 @@ import { HttpClientModule } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { NotificationsComponent } from './notifications/notifications.component';
+import { NotificationsComponent } from './notification-center/notifications/notifications.component';
 import { DashboardComponent } from './dashboard/dashboard.component';
+import { AlertsComponent } from './notification-center/alerts/alerts.component';
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
+import { DateStringPipe } from './pipes/dateString.pipe';
 
 @NgModule({
-  declarations: [	
+  declarations: [
     AppComponent,
     NotificationsComponent,
-      DashboardComponent
+    DashboardComponent,
+    AlertsComponent,
+    DateStringPipe
    ],
   imports: [
     BrowserModule,
     AppRoutingModule,
-    HttpClientModule
+    HttpClientModule,
+    BrowserAnimationsModule
   ],
   providers: [],
   bootstrap: [AppComponent]
